Module: start watcher only in dev mode

+++ modules/@hubblecommerce/hubble/module.js
@@ -179,43 +179,43 @@ export default async function (moduleOptions) {
         });
     });
 
-    const toTargetPath = (oldPath) => path.resolve(oldPath.replace(rootDir, targetDir));
+    if(this.options.dev) {
+        const toTargetPath = (oldPath) => path.resolve(oldPath.replace(rootDir, targetDir));
+        const excludedDirectories = [...dirBlacklist.map((__blacklistedDir) => `${rootDir}/${__blacklistedDir}/**`)];
+        chokidar.watch(`${rootDir}`, { ignoreInitial: true, ignored: excludedDirectories })
+            .on('all', async (event, filePath) => {
+                let newDestination = toTargetPath(filePath);
 
-    const excludedDirectories = [...dirBlacklist.map((__blacklistedDir) => `${rootDir}/${__blacklistedDir}/**`)];
-
-    chokidar.watch(`${rootDir}`, { ignoreInitial: true, ignored: excludedDirectories })
-        .on('all', async (event, filePath) => {
-            let newDestination = toTargetPath(filePath);
-
-            const hasApiSpecificSubfolders = apiTypeDirs.filter((__apiTypeDir) => filePath.includes(__apiTypeDir));
-            if (hasApiSpecificSubfolders.length !== 0) {
-                if (filePath.includes(`/${process.env.API_TYPE}/`)) {
-                    newDestination = toTargetPath(filePath.replace(`/${process.env.API_TYPE}/`, '/'));
-                } else {
-                    return;
-                }
-            }
-
-            if (event === 'add' || event === 'change') {
-                await fse.copy(filePath, newDestination);
-            }
-
-            if (event === 'unlink') {
-                const modulePath = filePath.replace(rootDir, baseDir);
-
-                fse.pathExists(modulePath, async (err, exists) => {
-                    if (exists) {
-                        // copy from module
-                        await fse.copy(modulePath, newDestination);
-                    } else if (!exists) {
-                        // path does not exist in module just remove from srcDir
-                        await fse.remove(newDestination);
-                    } else if (err) {
-                        console.log("err occurred: ", err);
+                const hasApiSpecificSubfolders = apiTypeDirs.filter((__apiTypeDir) => filePath.includes(__apiTypeDir));
+                if (hasApiSpecificSubfolders.length !== 0) {
+                    if (filePath.includes(`/${process.env.API_TYPE}/`)) {
+                        newDestination = toTargetPath(filePath.replace(`/${process.env.API_TYPE}/`, '/'));
+                    } else {
+                        return;
                     }
-                });
-            }
-        });
+                }
+
+                if (event === 'add' || event === 'change') {
+                    await fse.copy(filePath, newDestination);
+                }
+
+                if (event === 'unlink') {
+                    const modulePath = filePath.replace(rootDir, baseDir);
+
+                    fse.pathExists(modulePath, async (err, exists) => {
+                        if (exists) {
+                            // copy from module
+                            await fse.copy(modulePath, newDestination);
+                        } else if (!exists) {
+                            // path does not exist in module just remove from srcDir
+                            await fse.remove(newDestination);
+                        } else if (err) {
+                            console.log("err occurred: ", err);
+                        }
+                    });
+                }
+            });
+    }
 }
 
 // avoid registering the same module twice
